chore: split iDevice controls into multiple files
